fix(reward): change accrual_rate to 2byte

+++ programs/restaking/src/events/user_deposited_to_fund.rs
@@ -13,7 +13,7 @@ pub struct UserDepositedToFund {
     pub user_supported_token_account: Option<Pubkey>,
 
     pub wallet_provider: Option<String>,
-    pub contribution_accrual_rate: Option<u8>, // 100 is 1.0
+    pub contribution_accrual_rate: Option<u16>, // 100 is 1.0
     pub deposited_amount: u64,
     pub minted_receipt_token_amount: u64,
 }
