+++ programs/restaking/src/modules/fund/deposit_metadata.rs
@@ -7,7 +7,7 @@ use crate::modules::ed25519;
 pub struct DepositMetadata {
     user: Pubkey,
     wallet_provider: String,
-    contribution_accrual_rate: u8, // 100 is 1.0
+    contribution_accrual_rate: u16, // 100 is 1.0
     expired_at: i64,
 }
 
@@ -18,7 +18,7 @@ impl DepositMetadata {
         payload_signer_key: &Pubkey,
         user_key: &Pubkey,
         current_timestamp: i64,
-    ) -> Result<(String, u8)> {
+    ) -> Result<(String, u16)> {
         ed25519::SignatureVerificationService::verify(
             instructions_sysvar,
             self.try_to_vec()?.as_slice(),
